refactor: extract _hierarchy.py from _sections.py (14 functions, 3 constants)
Move hierarchy nesting, level normalisation, title-page stripping,
and collection title passes into a new _hierarchy.py module.
_sections.py drops from 1,955 to 1,207 lines.

Functions moved: _is_valid_rank_parent, _should_skip_same_keyword_nesting,
_respect_heading_rank_nesting, _collapse_degenerate_title_block,
_is_title_page_candidate, _strip_leading_title_page_sections,
_normalize_collection_titles, _broad_keywords_at_modal_rank,
_demote_same_rank_broad_keywords, _nest_broad_subdivisions,
_nest_chapters_under_broad_containers, _promote_more_prominent_heading_runs,
_flatten_single_work_title_wrapper, _equalize_orphan_level_gap.
